+++ www/src/extensions/query.test.ts
@@ -1,0 +1,13 @@
+import { describe, expect, it } from 'bun:test';
+
+import { queryExtension } from './query';
+
+describe('queryExtension', () => {
+  it('handles empty ranges', () => {
+    expect(queryExtension([])).toEqual([]);
+  });
+
+  it('handles query ranges', () => {
+    expect(() => queryExtension([{ from: 1, to: 2 }])).not.toThrow();
+  });
+});
